requirements.txt
streamlit
pandas
matplotlib.pyplot
matplotlib.ticker
seaborn